Play around with simpler boilerplate
Looks like we get a fair amount of inference, and we can avoid the
case let-ing for enum cases devoid of associated values.

Think we could take this further but wanted your thoughts:

    extension Routes
      enum iso {
        static let root = parenthesize <| PartialIso(const(.some(.root))) {
          $0 == .root ? unit : nil
        }

        static let pathComponents = parenthesize <| PartialIso(Routes.pathComponents) {
          guard case let .pathComponents(result) = $0 else { return nil }
          return result
        }
        …
      }
    }

Comes at a cost of losing the labels, though, which lend a lil clarity.
Just thought it might make the boilerplate smaller/easier.